Fix the issue with the right blade tilt motor turning on at boot time and never turning off.
rightBladeTilt0 is tied to GPIO5 which on the ESP32 is a boot-strapping
pin. This pin is tied to the output of an H-bridge driver that controls
the direction of the motor.

The problem is that an H‑bridge output is never guaranteed to be high‑impedance
at reset. Depending on the driver:

It may default LOW
It may default HIGH
It may actively drive both sides LOW (brake mode)
It may glitch during power‑up
It may briefly short the pin to ground through body diodes

The specific symptoms I had was that the right blade tilt motor would turn
on at boot time and never turn off. Even after connecting to an Xbox controller,
I was unable to control the motor.

While the best fix is to avoid using GPIO5 for this purpose, the quick
fix is to use a pull-up resistor on GPIO5. This will ensure that the ESP32
reads GPIO5 as high at reset, and thus boots into the default mode.

I accomplished this by setting GPIO5 to be a pull-up resistor very early
in the setup function before any other code runs. Then later you can set
it to output mode and use it as normal.